[EI-167] [web] test: Update import paths for mock handlers
