+++ web/__tests__/hooks/use-history-indicators-value-view-model.hook.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
-import { resetMockDB, mockDB } from '@/app/mocks/db.mock';
+import { resetMockDB, mockDB } from '@/app/mocks/db';
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import { useHistoryIndicatorsValueViewModel } from '@/app/business/hooks/indicator/use-history-indicators-value-view-model.hook';
 import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
