+++ web/__tests__/hooks/use-indicator-board-metadata-list-view-model.hook.test.tsx
@@ -7,7 +7,7 @@ import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import { act } from 'react-dom/test-utils';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
-import { resetMockDB } from '@/app/mocks/db.mock';
+import { resetMockDB } from '@/app/mocks/db';
 import { IndicatorBoardMetadata } from '@/app/business/services/view-model/indicator-board-metadata-view-model.service';
 
 const wrapper = SWRProviderWithoutCache;
